# [v7r2] Add callStack argument to S_ERROR to optimise CS