--- a/src/DIRAC/Core/Utilities/ReturnValues.py
+++ b/src/DIRAC/Core/Utilities/ReturnValues.py
@@ -14,14 +14,17 @@ import traceback
 from DIRAC.Core.Utilities.DErrno import strerror
 
 
-def S_ERROR(*args):
+def S_ERROR(*args, **kwargs):
   """ return value on error condition
 
   Arguments are either Errno and ErrorMessage or just ErrorMessage fro backward compatibility
 
   :param int errno: Error number
   :param string message: Error message
+  :param bool withStack: Set the CallStack attribute to an empty list for better performance
   """
+  withStack = kwargs.pop("withStack", True)
+
   result = {"OK": False, "Errno": 0, "Message": ""}
 
   message = ''
@@ -37,11 +40,13 @@ def S_ERROR(*args):
     message = "%s ( %s : %s)" % (strerror(result['Errno']), result['Errno'], message)
   result["Message"] = message
 
-  try:
-    callStack = traceback.format_stack()
-    callStack.pop()
-  except BaseException:
-    callStack = []
+  callStack = []
+  if withStack:
+    try:
+      callStack = traceback.format_stack()
+      callStack.pop()
+    except BaseException:
+      pass
 
   result["CallStack"] = callStack