When profiling some agents I've noticed that most of the time is spent generating the `CallStack` field in `S_ERROR`. This is due the CS being heavily used and even default keys have to format a full traceback. This PR adds a new `callStack` keyword argument to `S_ERROR` to avoid this which makes `gConfig.getOption` ~13x faster. Benchmarked with:

```python
import DIRAC
from DIRAC.Core.Base.Script import parseCommandLine
parseCommandLine()
DIRAC.gConfig.getOption("/Systems/Transformation/Production/Agents/WorkflowTaskAgent/PollingTim", 120)
%timeit DIRAC.gConfig.getOption("/Systems/Transformation/Production/Agents/WorkflowTaskAgent/PollingTim", 120)
```

Before:
```
137 µs ± 343 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

After:
```
11.3 µs ± 35.1 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```


BEGINRELEASENOTES

*Core
NEW: Add callStack keyword argument to S_ERROR to allow object to be created much faster

*Configuration
CHANGE: No longer include a full CallStack in the S_ERROR object returned by ConfigurationClient.getOption

ENDRELEASENOTES