@chaen 2021-06-07
I totally see the point, but how about instead of putting an empty list in the `callstack`, you just pass as argument `withStack=[ConfigurationClient.getOption]` ? This way it is very fast still, and yet we keep the information of the source at least ?