@chrisburr 2021-06-07
I like that idea a lot (except it should be `["ConfigurationClient.getOption"]` to keep it serialisable).

I've renamed `withStack=True` to be `callStack=None` that can be an override containing a list of strings.